should theoretically accept packets from the debug board
